Split out the remaining views- files that still need refactoring are missing docstrings at the top of file 3 Tests are commented out that need to be resolved
